Moved theano.sandbox.softsign to tensor/nnet/nnet.py. Added test.
