+++ theano/tensor/nnet/nnet.py
@@ -20,7 +20,7 @@ from six.moves import xrange
 import theano
 from theano import gof
 from theano import scalar
-from theano.tensor import basic as tensor, subtensor, opt
+from theano.tensor import basic as tensor, subtensor, opt, elemwise
 from theano.tensor.type import (values_eq_approx_remove_inf,
                                 values_eq_approx_remove_nan)
 from theano.tensor.opt import copy_stack_trace
@@ -2366,3 +2366,38 @@ def elu(x, alpha=1):
         Exponential Linear Units (ELUs)" <http://arxiv.org/abs/1511.07289>`.
     """
     return tensor.switch(x > 0, x, alpha * (tensor.exp(x) - 1))
+
+
+class ScalarSoftsign(theano.scalar.UnaryScalarOp):
+    """
+    Softsign activation function
+    :math:`\\varphi(\\mathbf{x}) = \\frac{1}{1+|x|}`
+
+    """
+    @staticmethod
+    def static_impl(x):
+        return x / (1.0 + abs(x))
+
+    def impl(self, x):
+        return ScalarSoftsign.static_impl(x)
+
+    def grad(self, inp, grads):
+        x, = inp
+        gz, = grads
+        if 'float' in x.type.dtype:
+            d = (1.0 + abs(x))
+            return [gz / (d * d)]
+        else:
+            return NotImplemented
+
+    def c_code(self, node, name, inp, out, sub):
+        x, = inp
+        z, = out
+        if node.inputs[0].type in [theano.scalar.float32,
+                                   theano.scalar.float64]:
+            return "%(z)s = %(x)s / (1.0+fabs(%(x)s));" % locals()
+        raise NotImplementedError('only floating point x is implemented')
+
+scalar_softsign = ScalarSoftsign(theano.scalar.upgrade_to_float,
+                                 name='scalar_softsign')
+softsign = elemwise.Elemwise(scalar_softsign, name='softsign')
